schema-toolkit/relaxng/src/main/java/eu/fox7/schematoolkit/relaxng/parser/NameProcessor.java
/**
 * Copyright 2009-2012 TU Dortmund
 *
 * This file is part of FoXLib.
 *
 * FoXLib is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Lesser General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * FoXLib is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with BonXai.  If not, see <http://www.gnu.org/licenses/>.
 */

package eu.fox7.schematoolkit.relaxng.parser;

import eu.fox7.schematoolkit.relaxng.om.Name;
import eu.fox7.schematoolkit.relaxng.om.RelaxNGSchema;
import eu.fox7.schematoolkit.relaxng.parser.exceptions.InvalidQNameException;
import eu.fox7.schematoolkit.xsd.tools.NameChecker;

import org.w3c.dom.Attr;
import org.w3c.dom.NamedNodeMap;
import org.w3c.dom.Node;

/**
 * This processor handles the name elements of a Relax NG XSDSchema
 *
 * @author Lars Schmidt
 */
class NameProcessor extends RNGProcessorBase {

    public NameProcessor(RelaxNGSchema rngSchema) {
        super(rngSchema, null);
    }

    @Override
    protected Name processNode(Node node) throws Exception {

        /**
         * Attributes:
         * -----------
         * ns String
         */
        String nameContent = node.getTextContent();
        if (nameContent == null || !NameChecker.isQName(nameContent)) {
            throw new InvalidQNameException(nameContent, "content of name");
        }

        if (getDebug()) {
            System.out.println("name: \"" + nameContent + "\"");
        }

        Name name = new Name(nameContent);

        NamedNodeMap attributes = node.getAttributes();
        if (attributes != null && attributes.getNamedItem("ns") != null) {
            String nsName = ((Attr) attributes.getNamedItem("ns")).getValue();
            name.setAttributeNamespace(nsName);
        }
        return name;
    }

    @Override
    protected void processChild(Node childNode) throws Exception {
        /**
         * Children:
         * ---------
         * (#PCDATA) NCName as content
         * 
         */
    }
}
